requirements.txt
flask
google-generativeai
os